# open-api-definition.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/trainings": {
      "post": {
        "tags": [
          "Trainings"
        ],
        "summary": "Create a new training",
        "description": "Creates a new training with the provided details.",
        "operationId": "createTraining",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/trainings/search": {
      "post": {
        "tags": [
          "Trainings"
        ],
        "summary": "Search trainings",
        "description": "Searches for trainings based on criteria.",
        "operationId": "searchTrainings",
        "parameters": [
          {
            "name": "arg0",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingSearchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTrainingDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trainings/{id}": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "summary": "Get training details",
        "description": "Retrieves details of a training by ID.",
        "operationId": "getTrainingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TrainingType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "TrainingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "trainerUsername": {
            "type": "string",
            "description": "The username of the trainer",
            "example": "john.doe"
          },
          "traineeUsername": {
            "type": "string",
            "description": "The username of the trainee",
            "example": "jane_doe"
          },
          "name": {
            "type": "string",
            "description": "The name of the training",
            "example": "Leg day",
            "maxLength": 50,
            "minLength": 1
          },
          "type": {
            "$ref": "#/components/schemas/TrainingType",
            "description": "The type of the training",
            "example": "{name:YOGA, id:1}"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "The date of the training",
            "example": "2021-12-31T23:59:59"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int64",
            "description": "The duration of the training in minutes",
            "example": 60,
            "minimum": 1
          }
        },
        "required": [
          "date",
          "durationMinutes",
          "name",
          "traineeUsername",
          "trainerUsername",
          "type"
        ]
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TrainingSearchDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "john.doe"
          },
          "fromDate": {
            "type": "string",
            "format": "date-time",
            "description": "The first name of the user",
            "example": "john"
          },
          "toDate": {
            "type": "string",
            "format": "date-time",
            "description": "The last name of the user",
            "example": "doe"
          },
          "isTrainee": {
            "type": "boolean",
            "description": "The active state of the user",
            "example": true
          }
        }
      },
      "PageTrainingDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/Sortnull"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "pageable": {
            "$ref": "#/components/schemas/Pageablenull"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "Pageablenull": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/Sortnull"
          },
          "paged": {
            "type": "boolean"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "Sortnull": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}